+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Enhancements:
 
+- Move `aim reindex` command under `aim storage` group (mihran113)
 - Add the ability to attach/remove tags on the Run Page (roubkar)
 - Support dictionary as an argument of `Run.track` (alberttorosyan)
 - Display the tags of the run in the tables of the explorers (VkoHov)